% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stukupuku.R
\name{stukupuku}
\alias{stukupuku}
\title{stukupuku}
\usage{
stukupuku(i = NULL, color = "red")
}
\arguments{
\item{i}{index of stukupuku line. By default it's random}

\item{color}{color of printed text. By default it's red}
}
\description{
returns one of stukupuku line
}
