% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stukupukuVec.R
\name{stukupukuVec}
\alias{stukupukuVec}
\title{stukupukuVec}
\usage{
stukupukuVec(v = c(1), color = "red")
}
\arguments{
\item{v}{vector of indices of stukupuku lines. By default it's first line}

\item{color}{color of printed text. By default it's red}
}
\description{
returns subseries of stukupuku lines
}
